requirements_recon.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.2.0
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
cachetools==5.3.3
certifi @ file:///croot/certifi_1707229174982/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
contourpy==1.1.1
cycler==0.12.1
einops==0.7.0
fastmri==0.3.0
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
google-auth==2.28.1
google-auth-oauthlib==1.0.0
grpcio==1.62.0
h5py==3.11.0
idna @ file:///croot/idna_1666125576474/work
imageio==2.34.0
importlib-metadata==7.0.1
importlib-resources==6.4.5
kiwisolver==1.4.5
lazy-loader==0.3
lightning-utilities==0.10.1
lpips==0.1.4
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.6.0
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
multidict==6.0.5
networkx==3.1
nibabel==5.2.1
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==23.2
pandas==2.0.3
Pillow==9.5.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyDeprecate==0.3.1
pyparsing==3.1.2
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
pytorch-lightning==1.4.2
pytorch-msssim==1.0.0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0
requests @ file:///croot/requests_1707355572290/work
requests-oauthlib==1.3.1
rsa==4.9
runstats==2.0.0
scikit-image==0.21.0
scipy==1.10.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
taming-transformers==0.0.1
taming-transformers-rom1504==0.0.6
tensorboard==2.14.0
tensorboard-data-server==0.7.2
test-tube==0.7.5
tifffile==2023.7.10
torch==1.11.0
torch-fidelity==0.3.0
torchmetrics==0.7.0
torchvision==0.12.0
tqdm==4.64.1
typing-extensions @ file:///croot/typing_extensions_1705599297034/work
tzdata==2025.2
urllib3 @ file:///croot/urllib3_1707770551213/work
werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
